// src/app/models/user.model.ts
export interface User {
    customer_id: number;
    name: String;
    email: String;
    password: String;
    credit_card: String;
    address_1: String;
    address_2: String;
    city: String;
    region: String;
    postal_code: String;
    country: String;
    shipping_region_id: number;
    day_phone: number;
    eve_phone: number;
    mob_phone: number;
}